vacuum-unsup/api/requirements.txt
numpy
opencv-python
fastapi
uvicorn
efficientnet_pytorch
torchvision
pydantic
Pillow
typing
torch